create paymethod by payos success

+++ JavaProjectBE/order-service/src/main/java/org/jewelryshop/orderservice/dto/response/OrderDetailResponse.java
@@ -13,6 +13,6 @@ public class OrderDetailResponse {
     String orderId;
     private String productId;
     private int quantity;
-    private Double pricePerUnit;
-    private Double totalPrice;
+    private int pricePerUnit;
+    private int totalPrice;
 }
