+++ JavaProjectBE/product-service/src/main/java/org/jewelryshop/productservice/dto/request/ProductRequest.java
@@ -10,8 +10,8 @@ import lombok.*;
 public class ProductRequest {
     private String name;
     private String description;
-    private double price;
-    private double originalPrice;
+    private int price;
+    private int originalPrice;
     private int stockQuantity;
     private String categoryId;
     private String brandId;
